实现整体数据管理的重新设计，制作简单的阵营UI，并且为阵营制作武器槽，当前武器可发射子弹，子弹可伤害指定敌方阵营

+++ Assets/_My/Scripts/UI/Panel/AnimationPanel/AnimationPanel.cs
@@ -18,7 +18,7 @@ namespace Slap.UI
 			
 			if (giftAnimationData.type == GiftAnimationData.AnimationType.windowed)
 			{
-				if (playerData.userCamp == 1)
+				if ((int)playerData.userCamp == 1)
 				{
 					LeftAnimation.gameObject.SetActive(true);
 					LeftAnimation.Play(giftAnimationData.AnimationName);
